Reduce the capcity of ring buffer
It seems that a large buffer can allow incessant entries
for an event which happens frequetly(e.g. `scheduler_tick()`).
This block our ctrlc handler from terminating the process. A
workaround is reducing the buffer capcity.

The better way could be providing ring_buffer__consume_n which
is availible on C interface.